[PopulationLink] Move functionality to util
